project tags merged with project details
