+++ tools/sense_studio/project_tags.py
@@ -2,9 +2,8 @@ import urllib
 
 from flask import Blueprint
 from flask import jsonify
-from flask import render_template
-from flask import request
 from flask import redirect
+from flask import request
 from flask import url_for
 
 from tools.sense_studio import project_utils
@@ -12,22 +11,11 @@ from tools.sense_studio import project_utils
 project_tags_bp = Blueprint('project_tags_bp', __name__)
 
 
-@project_tags_bp.route('/<string:project>', methods=['GET'])
-def tags_page(project):
+@project_tags_bp.route('/create-project-tag/<string:project>', methods=['POST'])
+def create_tag_in_project_tags(project):
+    data = request.form
     project = urllib.parse.unquote(project)
     path = project_utils.lookup_project_path(project)
-    config = project_utils.load_project_config(path)
-    project_tags = config.get('project_tags', {})
-    project_tags = {v: k for k, v in project_tags.items()}
-
-    return render_template('project_tags.html', path=path, project=project, project_tags=project_tags)
-
-
-@project_tags_bp.route('/create-project-tag', methods=['POST'])
-def create_tag_in_project_tags():
-    data = request.form
-    project = data['project']
-    path = data['path']
     config = project_utils.load_project_config(path)
     tag_name = data['tag']
 
@@ -40,13 +28,14 @@ def create_tag_in_project_tags():
 
     config['project_tags'] = project_tags
     project_utils.write_project_config(path, config)
-    return redirect(url_for('project_tags_bp.tags_page', project=project))
+    return redirect(url_for("project_details", project=project))
 
 
-@project_tags_bp.route('/remove-project-tag', methods=['POST'])
-def remove_tag_from_project_tags():
+@project_tags_bp.route('/remove-project-tag/<string:project>', methods=['POST'])
+def remove_tag_from_project_tags(project):
     data = request.json
-    path = data['path']
+    project = urllib.parse.unquote(project)
+    path = project_utils.lookup_project_path(project)
     remove_tag_idx = int(data['tagIdx'])
     config = project_utils.load_project_config(path)
     project_tags = config['project_tags']
@@ -65,10 +54,11 @@ def remove_tag_from_project_tags():
     return jsonify(success=True)
 
 
-@project_tags_bp.route('/edit-project-tag', methods=['POST'])
-def edit_tag_in_project_tags():
+@project_tags_bp.route('/edit-project-tag/<string:project>', methods=['POST'])
+def edit_tag_in_project_tags(project):
     data = request.json
-    path = data['path']
+    project = urllib.parse.unquote(project)
+    path = project_utils.lookup_project_path(project)
     tag_idx = data['tagIdx']
     new_tag_name = data['newTagName']
     config = project_utils.load_project_config(path)
