+++ tools/sense_studio/static/main.js
@@ -402,3 +402,88 @@ async function deselectTagFromList(idx, tagIndex, tagName, path, className) {
         selectTagDropdown.insertAdjacentHTML('beforeend', tagOption);
     }
 }
+
+///////////////////////////////////////////// Project Tags Operations //////////////////////////////////////////////////
+
+function checkIfTagExist(projectTags, tagId, errorLabelId, tagOperation) {
+    let tag = document.getElementById(tagId);
+    let errorLabel = document.getElementById(errorLabelId);
+    let tagButton = document.getElementById(tagOperation);
+    projectTagNames = Object.values(projectTags);
+
+    let disabled = false;
+
+    if (tag.value === '') {
+        setFormWarning(errorLabel, tag, '');
+        disabled = true;
+    } else if (projectTagNames.includes(tag.value)) {
+        setFormWarning(errorLabel, tag, 'This tag name already exist');
+        disabled = true;
+    } else {
+         setFormWarning(errorLabel, tag, '');
+    }
+
+    addTag.disabled = disabled;
+}
+
+
+function editProjectTag(tagIdx) {
+    let tagShow = document.getElementById(`tagShow${tagIdx}`);
+    let tagEdit = document.getElementById(`tagEdit${tagIdx}`);
+
+    tagEdit.classList.remove('uk-hidden');
+    tagShow.classList.add('uk-hidden');
+}
+
+
+async function saveProjectTag(tagIdx, url, projectTags) {
+    let tag = document.getElementById(`tag${tagIdx}`);
+    let tagShow = document.getElementById(`tagShow${tagIdx}`);
+    let tagEdit = document.getElementById(`tagEdit${tagIdx}`);
+    let path = document.getElementById('path').value;
+    let project = document.getElementById('projectName').value;
+
+    data = {
+        path: path,
+        tagIdx: tagIdx,
+        newTagName: tag.value,
+    };
+
+    let response = await asyncRequest(url, data);
+
+    if (response.success) {
+        tagEdit.classList.add('uk-hidden');
+        tagShow.classList.remove('uk-hidden');
+        window.location.href = `/project/${project}`;
+    }
+}
+
+
+async function removeProjectTag(tagIdx, url) {
+    let project = document.getElementById('projectName').value;
+    let path = document.getElementById('path').value;
+    data = {
+        path: path,
+        tagIdx: tagIdx,
+    };
+
+    let response = await asyncRequest(url, data);
+    if (response.success) {
+        window.location.href = `/project/${project}`;
+    }
+}
+
+
+function cancelEditProjectTag(tagIdx) {
+    let tagShow = document.getElementById(`tagShow${tagIdx}`);
+    let tagEdit = document.getElementById(`tagEdit${tagIdx}`);
+    let tag = document.getElementById(`tag${tagIdx}`);
+    let project = document.getElementById('projectName').value;
+    let error = document.getElementById('error');
+
+    setFormWarning(error, tag, '');
+
+    tag.value = tag.attributes.value.value;
+    tagEdit.classList.add('uk-hidden');
+    tagShow.classList.remove('uk-hidden');
+}
