+++ tools/sense_studio/templates/project_details.html
@@ -45,6 +45,7 @@
     </div>
 
     <input type="hidden" id="projectName" value="{{ project }}">
+    <input type="hidden" id="path" value="{{ path }}">
 
     <div class="uk-margin-medium-top" uk-alert>
         <a class="uk-alert-close" uk-close></a>
@@ -61,6 +62,89 @@
                     certain actions.
                 </div>
             </div>
+        </div>
+    </div>
+
+    <div class="uk-margin-medium-top uk-margin-medium-bottom temporal {{ 'uk-hidden' if not config.temporal }}" uk-grid>
+        <div>
+            <h1>Tags</h1>
+        </div>
+
+        <div>
+            <button class="uk-button uk-button-default uk-width-1-1" type="submit">
+                <span uk-icon="icon: plus"></span>
+                Create New Tag
+            </button>
+            <div class="uk-width-large" uk-drop="pos: bottom-center; mode:click">
+                <form class="uk-card uk-card-default uk-card-hover uk-form-stacked" method="POST"
+                      action="{{ url_for('project_tags_bp.create_tag_in_project_tags', project=project) }}">
+                    <div class="uk-card-body">
+                        <div>
+                            <label class="uk-form-label uk-text-danger" id="tagLabel"></label>
+                            <div class="uk-inline uk-width-1-1 uk-margin-bottom">
+                                <span class="uk-form-icon" uk-icon="icon: tag"></span>
+                                <input class="uk-input" type="text" id="tag" name="tag" placeholder="Tag Name"
+                                       oninput="checkIfTagExist({{ project_tags }}, 'tag', 'tagLabel', 'addTag');" autocomplete="off">
+                            </div>
+                        </div>
+
+                        <button class="uk-button uk-button-default uk-width-1-1" type="submit" id="addTag" disabled>
+                            <span uk-icon="icon: plus"></span>
+                            Add
+                        </button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    </div>
+
+    <div class="temporal {{ 'uk-hidden' if not config.temporal }}">
+        <h5 id="error" class="uk-text-danger uk-text-center"></h5>
+        <div class="uk-child-width-1-1@m uk-child-width-1-3@l uk-grid-small uk-grid-match" uk-grid>
+            {% for tag_idx, tag_name in project_tags.items() %}
+                <div class="uk-card uk-card-default uk-card-header">
+                    <div id="tagShow{{ tag_idx }}" uk-grid>
+                        <div class="uk-width-expand">
+                            <h3 class="uk-card-title uk-float-left">{{ tag_name }}</h3>
+                        </div>
+                        <div class="uk-width-auto">
+                            <a uk-icon="icon: pencil; ratio: 1.1" uk-tooltip="Edit Tag" onclick="editProjectTag('{{ tag_idx }}');"></a>
+                       </div>
+                        <div class="uk-width-auto">
+                            <a href="#" uk-icon="icon: trash; ratio: 1.1" uk-tooltip="Remove Tag"></a>
+                            <div uk-dropdown="mode: click; pos:bottom-justify">
+                                <p>
+                                    Remove tag from the project tags in the config.
+                                </p>
+                                <a class="uk-button uk-button-danger"
+                                    onclick="removeProjectTag('{{ tag_idx }}', '{{ url_for('project_tags_bp.remove_tag_from_project_tags', project=project) }}');">
+                                    <span uk-icon="icon: trash"></span> REMOVE
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+
+                    <div id="tagEdit{{ tag_idx }}" class="uk-hidden">
+                        <form>
+                            <div class="uk-form-control">
+                                <div class="uk-inline uk-width-1-1 uk-margin-bottom">
+                                    <span class="uk-form-icon" uk-icon="icon: tag"></span>
+                                    <input class="uk-input" type="text" placeholder="Tag name" autocomplete="off"
+                                           oninput="checkIfTagExist({{ project_tags }}, 'tag{{ tag_idx }}', 'error', 'saveTag{{ tag_idx }}');" id="tag{{ tag_idx }}"
+                                           value="{{ tag_name }}" autocomplete="off">
+                                </div>
+                            </div>
+                            <a class="uk-button uk-button-primary" uk-tooltip="Save Tag" id="saveTag{{ tag_idx }}"
+                               onclick="saveProjectTag('{{ tag_idx }}', '{{ url_for('project_tags_bp.edit_tag_in_project_tags', project=project) }}', {{ project_tags }});">
+                                <span uk-icon="icon: check"></span>
+                            </a>
+                            <a class="uk-button uk-button-danger" uk-tooltip="Cancel" onclick="cancelEditProjectTag('{{ tag_idx }}');">
+                                <span uk-icon="icon: close"></span>
+                            </a>
+                        </form>
+                    </div>
+                </div>
+            {% endfor %}
         </div>
     </div>
 
@@ -94,19 +178,6 @@
                     </div>
                 </form>
             </div>
-        </div>
-
-        <div class="temporal {{ 'uk-hidden' if not config.temporal }}">
-            <h1>Tags</h1>
-        </div>
-
-        <div class="temporal {{ 'uk-hidden' if not config.temporal }}">
-            <form method="get" action="{{ url_for('project_tags_bp.tags_page', project=project) }}">
-                <button class="uk-button uk-button-default uk-width-1-1" type="submit">
-                    <span uk-icon="icon: plus"></span>
-                    Create New Tags
-                </button>
-            </form>
         </div>
     </div>
