--- tools/sense_studio/templates/project_tags.html
@@ -1,88 +1,0 @@
-{% extends 'skeleton.html' %}
-
-{% block title %} Assign Tags {% endblock %}
-
-{% block navigation %}
-    {% include 'navigation.html' %}
-{% endblock navigation %}
-
-{% block main %}
-    <div class="uk-container uk-container-xsmall">
-        <h1 class="uk-heading-medium uk-margin-large-top uk-margin-remove-bottom">Tags</h1>
-         <div class="uk-text-meta uk-text-large uk-margin-medium-bottom">
-            <span uk-icon="icon: folder"></span> {{ path }}
-        </div>
-
-        <form class="uk-form-stacked" method="POST" action="{{ url_for('.create_tag_in_project_tags') }}"
-              onsubmit="return createProjectTag({{ project_tags }});">
-            <input type="hidden" id="project" name="project" value="{{ project }}">
-            <input type="hidden" id="path" name="path" value="{{ path }}">
-
-            <div class="uk-child-width-1-1 uk-grid-small" uk-grid>
-                <div>
-                    <label class="uk-form-label" for="tag">Tag name</label>
-                    <label class="uk-form-label uk-text-danger" id="tagLabel"></label>
-                    <div class="uk-form-controls uk-width-1-1">
-                        <input class="uk-input" type="text" id="tag" name="tag" placeholder="Tag name" autocomplete="off" oninput="return checkIfTagExist({{ project_tags }}, 'tag', 'tagLabel');">
-                    </div>
-                </div>
-                <div>
-                    <button type="submit" class="uk-button uk-button-primary" uk-tooltip="Add Tag"><span uk-icon="icon: plus"></span> Add</button>
-                </div>
-            </div>
-        </form>
-
-        <h3 class="uk-heading-line uk-text-center"><span>Project Tags</span></h3>
-        <h5 id="error" class="uk-text-danger uk-text-center"></h5>
-        <form>
-            <div class="uk-child-width-1-1@m uk-child-width-1-2@l uk-grid-small uk-grid-match" uk-grid>
-                {% for tag_idx, tag_name in project_tags.items() %}
-                    <div class="uk-card uk-card-default uk-card-header">
-                        <div id="tagEdit{{ tag_idx }}" class="uk-hidden">
-                            <div class="uk-form-control">
-                                <div class="uk-inline uk-width-1-1 uk-margin-bottom">
-                                    <span class="uk-form-icon" uk-icon="icon: tag"></span>
-                                    <input class="uk-input" type="text" placeholder="Tag name" autocomplete="off"
-                                           oninput="return checkIfTagExist({{ project_tags }}, 'tag_{{ tag_idx }}', 'error');" id="tag_{{ tag_idx }}"
-                                           value="{{ tag_name }}" autocomplete="off">
-                                </div>
-                            </div>
-                            <a class="uk-button uk-button-primary" uk-tooltip="Save Tag"
-                               onclick="saveProjectTag('{{ tag_idx }}', '{{ url_for('.edit_tag_in_project_tags') }}', {{ project_tags }});">
-                                <span uk-icon="icon: check"></span>
-                            </a>
-                            <a class="uk-button uk-button-danger" uk-tooltip="Cancel" onclick="cancelEditProjectTag('{{ tag_idx }}');">
-                                <span uk-icon="icon: close"></span>
-                            </a>
-                        </div>
-
-                        <div id="tagShow{{ tag_idx }}" uk-grid>
-                            <div class="uk-width-expand">
-                                <h3 class="uk-card-title uk-float-left">{{ tag_name }}</h3>
-                            </div>
-                            <div class="uk-width-auto">
-                                <a uk-icon="icon: file-edit" uk-tooltip="Edit Tag" onclick="editProjectTag('{{ tag_idx }}');"></a>
-                           </div>
-                            <div class="uk-width-auto">
-                                <a href="#" uk-icon="icon: trash; ratio: 1.1" uk-tooltip="Remove Tag"></a>
-                                <div uk-dropdown="mode: click; pos:bottom-justify">
-                                    <p>
-                                        Remove tag from the project tags in the config.
-                                    </p>
-                                    <a class="uk-button uk-button-danger"
-                                        onclick="removeProjectTag('{{ tag_idx }}', '{{ url_for('.remove_tag_from_project_tags') }}');">
-                                        <span uk-icon="icon: trash"></span> REMOVE
-                                    </a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                {% endfor %}
-            </div>
-        </form>
-    </div>
-{% endblock %}
-
-{% block scripts %}
-    <script src="{{ url_for('static', filename='project_tags.js') }}"></script>
-{% endblock %}
